App now uses TwitterColorEmoji font for emojis, added About page
I've been looking for a OS10-compatible font with twitter-style emoji as the system one is very old and many codepoints are missing.. I randomly stumbled upon Twemoji.ttf when looking at the official Twitter apk, and it works with a bit of workaround: 
It needs to be used ONLY for emojis, since it seems to break any font priority for other characters. So now tweets are displayed with labels in Html mode, and every tweet is parsed char by char, adding around emojis suitable <span> elements that set the font to Twemoji only for those characters.
It's not very efficient but the app performance is still good (luckily there are at most 180 characters per tweet!)