Dynamic User-Agent download to keep it up-to-date
Previously the user agent used in api requests was hardcoded, and when it became obsolete it would cause issues with requests not being handled correctly by the server (trends api had most issues). Now it downloads the latest u.a. from the web and uses a hardcoded value as fallback